+++ fastly/ngwaf/v1/events/api_response.go
@@ -1,6 +1,11 @@
 package events
 
-import "time"
+import (
+	"time"
+
+	"github.com/fastly/go-fastly/v10/fastly/ngwaf/v1/requests"
+	"github.com/fastly/go-fastly/v10/fastly/ngwaf/v1/signals"
+)
 
 // Event is the API response structure for the get, list, and expire operations.
 type Event struct {
@@ -25,13 +30,13 @@ type Event struct {
 	// IsExpired if true, the event should be set to expired.
 	IsExpired bool `json:"is_expired"`
 	// Reasons is a list of signals and their counts.
-	Reasons []Reason `json:"reasons"`
+	Reasons []signals.Reason `json:"reasons"`
 	// RemoteHostname is the hostname of the event
 	RemoteHostname string `json:"remote_hostname"`
 	// RequestCount is the total numer of requests.
 	RequestCount int `json:"request_count"`
 	// SampleRequest is an example of a request that triggered the event.
-	SampleRequest map[string]any `json:"sample_request"`
+	SampleRequest requests.Request `json:"sample_request"`
 	// Source is the IP address of the source of the event.
 	Source string `json:"source"`
 	// Type is the type of event
@@ -40,14 +45,6 @@ type Event struct {
 	UserAgents []string `json:"user_agents"`
 	// Window is the time in seconds where the items were detected.
 	Window int `json:"window"`
-}
-
-// Reason is the reason an event was triggered.
-type Reason struct {
-	// Signal ID is the ID of the signal that triggered the event
-	SignalID string `json:"signal_id"`
-	// Count is the number of times this signal was detected
-	Count int `json:"count"`
 }
 
 // Events is the API response structure for the list events operation.