+++ fastly/ngwaf/v1/requests/doc.go
@@ -1,0 +1,2 @@
+// Package requests contains API operations to get all and specific requests made to a Fastly Next-Gen WAF.
+package requests